fix: call read orientation, not layout, inference (#39)
